+++ config/sandbox.ts
@@ -4,7 +4,7 @@ import {readValidators} from './utils';
 const staging: GlobalConfig = {
   ethereum: {
     contractRegistry: {
-      address: ''
+      address: '0x059FDd69e771645fe91d8E1040320DbB845cEaFd'
     },
     token: {
       totalSupply: '1000000' + '0'.repeat(18),
@@ -12,15 +12,16 @@ const staging: GlobalConfig = {
       symbol: 'UMB'
     },
     chain: {
-      padding: 60,
-      requiredSignatures: 1
+      padding: 300,
+      requiredSignatures: 1,
+      replicator: '0x620583c75bb474e06485893b795b0883b5816d10'
     },
     stakingBank: {
       minAmountForStake: 100n * BigInt(1e18),
     },
     validators: readValidators('https://validator.sbx.umb.network'),
   },
-  smartchain: {
+  bsc: {
     distributor: {
       address: '0x2BB2B0df0DB14EAfd5FCaBc46Fd543F7a6dFF8E9'
     },
@@ -33,8 +34,9 @@ const staging: GlobalConfig = {
       symbol: 'UMB'
     },
     chain: {
-      padding: 60,
-      requiredSignatures: 1
+      padding: 20,
+      requiredSignatures: 1,
+      replicator: '0x0760b24EC17dDf59893D7c16D75353dE24c00F14'
     },
     stakingBank: {
       minAmountForStake: 100n * BigInt(1e18),