+++ config/types.ts
@@ -19,6 +19,7 @@ export interface Config {
   chain: {
     padding: number
     requiredSignatures: number
+    replicator?: string
   },
   stakingBank: {
     minAmountForStake: bigint
@@ -28,5 +29,5 @@ export interface Config {
 
 export interface GlobalConfig {
   ethereum: Config,
-  smartchain: Config,
+  bsc: Config,
 }