+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.0.0] - 2021-10-04
+### Added
+- Foreign Chain
+- `Registry` contract has additional method `atomicUpdate`
+
+### Changed
+- redeployment process for `ForeignChain` is different from `Chain`
+
 ## [3.2.0] - 2021-08-31
 ### Changed
 - turn off power until we have proper DPoS